pass a channel and buffer size rather than 3 separate variables
also make sure that the page size is a multiple of two by shrinking it
to the nearest power of two. This is a requirement by the perfmap
buffers.
